Fix #1731: Y.Calendar.selectDates fails when passed the maximumDate with minutes/seconds
Fixing the issue explained in #1731, which causes selectDates to fail when the user passes the maximumDate with minutes or seconds to it. The calendar code is already resetting the hours part of the date's time, but this diff resets everything to make sure that date selection doesn't take the date's time into account anymore.